requirements.txt
flask
gunicorn
python-dotenv
pydantic
google-generativeai
jinja2
starlette